+++ dependencies.php
@@ -11,7 +11,6 @@ declare(strict_types=1);
 
 namespace Flextype;
 
-use Slim\Flash\Messages;
 use Flextype\Component\I18n\I18n;
 use function Flextype\Component\I18n\__;
 
@@ -23,12 +22,6 @@ $flextype->registry->set('plugins.admin.settings.navigation.content.entries', ['
 $flextype->registry->set('plugins.admin.settings.navigation.extends.plugins', ['title' => __('admin_plugins'),'icon' => 'fas fa-plug', 'link' => $flextype->router->pathFor('admin.plugins.index')]);
 $flextype->registry->set('plugins.admin.settings.navigation.system.tools', ['title' => __('admin_tools'),'icon' => 'fas fa-toolbox', 'link' => $flextype->router->pathFor('admin.tools.index')]);
 $flextype->registry->set('plugins.admin.settings.navigation.system.api', ['title' => __('admin_api'),'icon' => 'fas fa-network-wired', 'link' => $flextype->router->pathFor('admin.api.index')]);
-
-// Add Global Vars Admin Twig Extension
-$flextype->twig->addExtension(new GlobalVarsAdminTwigExtension($flextype));
-
-// Add Flash Twig Extension
-$flextype->twig->addExtension(new FlashTwigExtension($flextype));
 
 /**
  * Add Assets
@@ -45,13 +38,6 @@ $flextype['registry']->set('assets.admin.js',
                        array_merge($_admin_js,
                        ['project/plugins/admin/assets/dist/js/admin-vendor-build.min.js']));
 
-/**
- * Add flash service to Flextype container
- */
-$flextype['flash'] = static function () {
-    return new Messages();
-};
-
 $flextype['DashboardController'] = static function ($container) {
     return new DashboardController($container);
 };
@@ -66,10 +52,6 @@ $flextype['PluginsController'] = static function ($container) {
 
 $flextype['EntriesController'] = static function ($container) {
     return new EntriesController($container);
-};
-
-$flextype['UsersController'] = static function ($container) {
-    return new UsersController($container);
 };
 
 $flextype['ToolsController'] = static function ($container) {